refactor: cgi response 처리 재구성을 위한 cgi::Responder 클래스 구현
