+++ src/handler/builder/ErrorBuilder.cpp
@@ -1,7 +1,7 @@
 // ErrorBuilder.cpp
 #include "ErrorBuilder.hpp"
 
-#include "../response/Response.hpp"
+#include "../utils/response.hpp"
 
 using namespace handler::builder;
 
@@ -17,7 +17,7 @@ std::string ErrorBuilder::joinAllowMethods(const std::vector<std::string>& metho
 
 http::Packet ErrorBuilder::build(const router::RouteDecision& decision, const http::Packet&,
 								 const config::Config& config) const {
-	http::Packet response = response::makeErrorResponse(
+	http::Packet response = utils::makeErrorResponse(
 		decision.status, &config, http::StatusCode::to_reasonPhrase(decision.status),
 		http::ContentType::to_string(http::ContentType::CONTENT_TEXT_PLAIN));
 	std::string allow = joinAllowMethods(decision.allowMethods);
